Integrate neroxis toolsuite (#60)
This adds some buttons to create the necessary textures for the new
shaders.
Under the hood this integrates the neroxis toolsuite, a CLI program
written in Java. That's why you need to specify a java environment.
Luckily the FAF client also runs on java, so if you have that installed
there is already a java environment that you can use.

This PR also removes the two neroxis buttons that you could use to
generate a heightmap. The used neroxis version was way outdated and we
don't allow uploading maps based on the map generator to our map vault
anyway.