Improve Resource Browser UI (#62)
Adds an alpha channel preview and previews of the mipmaps at level 2, 4,
and 6.
Adds the texture resolution.
These changes only affect the layer textures. The resource browser is
still the same for decals and props. They just had to be made a little
bit smaller, so now 6 items of them fit in a row.
I also fixed the problem that the textures appeared upside down in the
layer tab.