More bugfixes (#71)
- The calculation of the z component of the normals could lead to
artifacts when it evaluated to 0. We now hide them by using the absolute
value instead.
- The terrain now preserves more details when zoomed out. The result
should be more similar to the experience in the game.
- When you use textures of different resolution in the Layers, they have
to be all scaled to the same size internally because they are stored in
an array texture. This scaling would lead to a recalculation of mipmaps,
so the blurred mipmaps would not be visible anymore, leading to a
different look than in the game. Mipmaps are now preserved properly.
- The skycube in the light settings now actually updates when you drag
in a texture. The skycube is only visible on unit reflections, so you
won't notice an immediate change.
- The sun direction on water reflections now updates properly
- The editor should now pause rendering when it is in the background,
saving resources
- The sun reflection setting in the water material was actually a
multiplier for the sun reflection amount. It has been removed because
its function was unclear and is now added back with a more descriptive
name.